section scrolling js aproach done
