+++ index.html
@@ -31,16 +31,30 @@
                 <a href="#" class="menu__link">Home</a>
               </li>
               <li class="menu__item">
-                <a href="#" class="menu__link">About us</a>
+                <a
+                  href="javascript:void(0)"
+                  id="about-us-link"
+                  class="menu__link"
+                  >About us</a
+                >
               </li>
               <li class="menu__item">
-                <a href="#" class="menu__link">Shop</a>
+                <a href="javascript:void(0)" id="shop-link" class="menu__link"
+                  >Shop</a
+                >
               </li>
               <li class="menu__item">
-                <a href="#" class="menu__link">Pages</a>
+                <a href="javascript:void(0)" id="pages-link" class="menu__link"
+                  >Pages</a
+                >
               </li>
               <li class="menu__item">
-                <a href="#" class="menu__link">Contact us</a>
+                <a
+                  href="javascript:void(0)"
+                  id="contact-us-link"
+                  class="menu__link"
+                  >Contact us</a
+                >
               </li>
             </ul>
             <div class="socials">
@@ -125,7 +139,7 @@
           </div>
         </div>
       </section>
-      <section class="about-section">
+      <section id="about-us" class="about-section">
         <div class="about__content section-content">
           <img
             src="/images/about-2x.png"
@@ -184,7 +198,7 @@
           </div>
         </div>
       </section>
-      <section class="products-section">
+      <section id="shop" class="products-section">
         <div class="products__content section-content">
           <p class="display-text products__display-text">Organic</p>
           <h2 class="heading-2 products__heading">Our Products</h2>
@@ -387,7 +401,7 @@
           </div>
         </div>
       </section>
-      <section class="testimonials">
+      <section id="pages" class="testimonials">
         <img
           src="/images/story-left-2x.jpg"
           alt="natural products as background image"
@@ -748,7 +762,7 @@
         </div>
       </section>
     </main>
-    <footer class="footer-section">
+    <footer id="contact-us" class="footer-section">
       <div class="footer__content section-content">
         <div class="footer__main">
           <div class="footer__contact-us">
