+++ public/styles/components/_socials.scss
@@ -1,4 +1,5 @@
 .socials {
+  width: 100%;
   display: grid;
   grid-template-columns: repeat(auto-fit, 6.4rem);
   gap: 2.4rem;
